Move app secter tests to android tests
